Added a workaround for the description feature
This adds the documentation of the class attributes inside the class if
the class attribute is an instance of `Description`.

I must say this is not the best approach, since it still feels starnge
to use.

Currently it looks like this and is a combination of the `var` approach
and the new typed approach.

    class Cls(TypeChecked):
        one: str = required("", """ One is the loneliest number""")

Compared to ealier

    class Cls:
        one = varString("", """ One is the loneliest number""", required=True)

it is atleast shorter.

What would have been nice if the attributes docstr would be available in
the python runtime via `.__doc__`, but sadly attribute docstr was
horseshoed in later and is only availbale in the AST. Then it could have
looked like this.

    class Cls(TypeChecked):
        one: Annotated[str, required()] = ""
        """ One is the loneliest number"""

And I would not need to use Metaclass magic.